+++ docstrfmt/__init__.py
@@ -9,4 +9,4 @@ from .const import (
 )
 from .docstrfmt import Manager
 
-__version__ = "2.1.1.dev0"
+__version__ = "2.1.2.dev0"